Add edit modue
